requirements.txt
asgiref==3.5.2
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
typing_extensions==4.4.0
django-autoslug==1.9.8
django-debug-toolbar==3.4.0